+++ README.md
@@ -1,6 +1,6 @@
 # com.utilities.audio
 
-[![openupm](https://img.shields.io/npm/v/com.utilities.audio?label=openupm&registry_uri=https://package.openupm.com)](https://openupm.com/packages/com.utilities.audio/)
+[![Discord](https://img.shields.io/discord/855294214065487932.svg?label=&logo=discord&logoColor=ffffff&color=7389D8&labelColor=6A7EC2)](https://discord.gg/xQgMW9ufN4) [![openupm](https://img.shields.io/npm/v/com.utilities.audio?label=openupm&registry_uri=https://package.openupm.com)](https://openupm.com/packages/com.utilities.audio/) [![openupm](https://img.shields.io/badge/dynamic/json?color=brightgreen&label=downloads&query=%24.downloads&suffix=%2Fmonth&url=https%3A%2F%2Fpackage.openupm.com%2Fdownloads%2Fpoint%2Flast-month%2Fcom.utilities.audio)](https://openupm.com/packages/com.utilities.audio/)
 
 A simple package for audio extensions and utilities in the [Unity](https://unity.com/) Game Engine.
 
@@ -56,4 +56,4 @@ A basic `AbstractRecordingBehaviour` is included in this package to make it very
 ## Audio Clip Extensions
 
 Provides extensions to encode `AudioClip`s to PCM encoded bytes.
-Supports 8 bit and 16 bit sample sizes.
+Supports 8, 16, 24, and 32 bit sample sizes.
